requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.7
Django==4.2.4
djangorestframework==3.14.0
Faker==19.3.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
